Made errors less painful, restructured Card
The Card restructure makes implementations of cards simpler as one doesn't need to hold onto an outer structure and unwrap it at every turn, while making implementation more obvious and allowing earlier failure during construction. Cons of this change are a less type-centric program flow around the type of card being handled and a dependency in the construction of a card. Pros are as listed and additionally removal of the Rc dependency